requirements.txt
langgraph>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
ipykernel>=6.0.0
urllib3<2
tavily-python>=0.5.0
tavily-python
